Move ImportConverter.
